Added navigation to the email
