+++ SignUp.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+import {Text} from 'react-native-elements';
+import {View} from 'react-native';
+
+
+export default class SignUp extends Component<{}> {
+
+    render() {
+        return (
+            <View>
+                <Text> For testing </Text>
+            </View>
+        )}
+}
+
+module.exports = SignUp;
