+++ package.json
@@ -11,9 +11,9 @@
     "react": "16.0.0",
     "react-native": "0.50.4",
     "react-native-elements": "^0.18.4",
+    "react-native-simple-auth": "^2.1.0",
     "react-native-vector-icons": "^4.4.2",
-    "react-native-simple-auth": "^2.1.0"
-
+    "react-navigation": "^1.0.0-beta.21"
   },
   "devDependencies": {
     "babel-jest": "21.2.0",
